shopify-cro-poc/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
google-generativeai>=0.8.0
python-dotenv>=1.0.0
